Remove outdated Django apps and related files
The commit has removed multiple Django applications such as 'Home', 'Usuarios', etc., and their related files. These removals include models, views, templates, and migration files. This is most likely due to a decision to refactor the logic of these apps into different or new Django apps. It is important to note that this would mean that all functionalities these apps were responsible for need to be implemented elsewhere in the project to maintain the same level of features in the application. Any integrated services, utilities, or applications depending on these apps may also need to be updated or refactored to accommodate for these changes.