Commit: Improve CSS styles and HTML template user interface
Refined the CSS and HTML files for a better User Interface (UI) experience. The CSS file 'styles.css' underwent several updates including the reordering and deleting of certain class definitions, and the addition of new ones to enhance performance and user comprehensibility.

Simultaneously, the HTML template 'ficha.html' has been updated for better User Experience (UX). The changes include a repositioning of elements, more organized presentation of data and simplified navigation. The updated files should improve the look and feel of the website making it more user-friendly and responsive.