Remove several migration files due to database overhaul
The previous migration files have been removed because of major changes in the database schema. The necessity of these changes is due to the added fields and relationships in the CustomUser and Ficha models, which resulted in a need to rebuild the database. The files that have been deleted are largely obsolete and are mostly associated with the email field in the CustomUser model and the status field in the Ficha model, etc. Also, a new html file has been added for displaying teacher's details. The models.py has also been updated reflecting the changes on the models.

Additionally, Keyword field for Ficha model has been changed to use TaggableManager for better management of tags. User Ficha relationship name has been updated. Now, in order to be more efficient, all previous migrations have been rolled into a single migration, hence, the previous files which are no longer needed, have been deleted.