Refactor CSS and Django templates; improve admin UI
The update involved multiple changes to both the Django templates, as well as the CSS. CSS revisions include removing unused classes and adding new ones to improve the design aesthetic. Modified Django templates so that the correct URL tags are used and adjusted the rendering of various forms as well as the layout of some pages. The Django admin interface was also enhanced for greater functionality and ease of use. The purpose of this refactor and update was to improve overall user experience and efficiency in managing backend data.