Add scalable 90% CSS class and .ease-out transition timing function
New CSS classes were added for scaling elements to 90% of their original size and implementing a cubic bezier as a transition timing function. Additionally, a new Django model was created to manage FichaImage objects, which allow for more flexible image handling in the context of a student's ficha. Django admin was also modified to handle these new objects.