Add 'anexos' field for embedded videos in Fichas
Introduced a new field 'anexos' in the Ficha model which allows embedded videos to be added with each Ficha. This change helps to improve visual content and potentially learner engagement. In line with this change, related Fichas' templates and forms were updated to include and handle this new field. Collection and presentation of these videos are handled with django-embed-video package, also added into requirements.txt.

Ensured the new field is correctly applied in the database by creating respective migrations files. Added required settings for the django-embed-video app in settings.py.

Lastly, relevant views were updated to process and present the embedded videos correctly.