Add Docker configurations for Django project
Added a Dockerfile and .dockerignore for the Django project. This sets up an isolated environment that runs the application with Python 3.11.4, Django, and other Python dependencies defined in requirements.txt, as well as Node.js and npm dependencies defined in package*.json. This improves the project setup process and application's portability because the environment is consistent, isolated, and easily replicated.

The Dockerfile also sets a few environment variables for Python and creates a dedicated directory in the container's filesystem (/code) for our application's source code. The Dockerfile ensures, using RUN and COPY commands, that all the required files are present and all the necessary installations are made in the container.

The .dockerignore file was added to prevent unnecessary or sensitive files from being copied into the Docker image, thus maintaining the image clean and secure.