Add ssh private and public keys to .dockerignore and .gitignore
Added ssh private ('ssh') and public ('ssh.pub') keys to .dockerignore and .gitignore files. These keys can potentially be sensitive information and hence are ignored from Docker and Git environments to prevent accidental upload of this information along with the project source code. It's a crucial step in maintaining the security of the project.