Introduce docker-compose for development and upgrade settings.py
A compose-dev.yaml file is added to utilize docker-compose for development. This change simplifies the setup and testing process for devs by removing the need to manually build and run Docker containers. The settings.py file in the FPPT16 project is also updated to use python-decouple for better handling of secret keys and environment variables. The 'django_extensions' is added to installed apps for debugging purposes and the 'NPM_BIN_PATH' is updated to use a system-agnostic path.