Set Spanish locale across the system and update Node.js version
Locale for Spanish has been added in `general_views.py` to display dates and time in the Spanish format. This change would enhance UX for Spanish users.

To ensure this change works across containers, Dockerfile has been updated to install the necessary locales and set the environment for Spanish.

Furthermore, the version of Node.js is updated to stay current with the LTS (Long-Term Support) version, which would enhance app performance and provide better security.