Remove locale set fallback, update locale set place and add dependency
Removed the redundant setlocale and Error pop up for fallback locale from `general_views.py`, this has been handled in a better way. Setlocale moved inside the function to load when locale-specific data needed.

Added Babel in `requirements.txt` which provides a way to get current date and time, based on Timezone. This usage of Babel will be much efficient and accurate than previous version.

The goal is to get the current date and time in Spain Language directly instead of setting the whole system locale to Spanish.